more tests and some fixes for WebdavPath
did more tests with WebdavPath, in particular for the edge/exception
conditions.
the tests are all in WebdavPathTest.java, some bug fixes in WebdavPath
for the edge coditions. There is an instance variable testwebdavclass

final boolean testwebdavclass = true;

leaving this as true does tests on WebdavPath,
setting it to false uses a uri for the "file://" based paths, returned
by system. this makes it feasible to compare between WebdavPath vs
"file://" implementations of path. Currently, only toAbsolutePath()
causes a failed test.

For absolute paths, they worked the same way, but for relative paths.
UnixPath patches it to "user.home" user's home directory, while in my
implementation, i patched it to path.getRoot() which if the root path is
"/" then the relative path is appended there

WebdavPath and UnixPath (internal to java) is rather similar in many
aspects.

However, some variances is discovered. Currently, WebdavPath
requires the other path to start with the current path and
that the current path is a subset of the other path.

"/a/b".relativize("/a/b/c/d") returns "c/d"

the current WebdavPath.relativize() returns the remaining subpath as a
relative path. This result conforms to JDK 1.8 javadoc spec.
works the same way in both UnixPath and WebdavPath.

however, UnixPath resolves the following

"/a/b/c/d".relativize("/a/b/e/f") returns "../../e/f"
"/a/b/c/d".relativize("/a/b") returns "../.."

WebdavPath throws IllegalArgumentException for these 'other' paths.
i.e. that used as the argument.
it isn't feasible currently to cater for the ".." based paths in
this simplified but rather complete implementation.
i.e. Webdavpath should work for most of the straightforward
absolute paths and relative paths appended on absolute paths.
But that it doesn't handles the more 'esoteric' ".." based resolutions.
This in part as there are complications when considering against root
paths etc. Those may be considered in future enhancements.

it should be pretty usable currently though.