new WebdavPath.toAbsolutePath(), updated javadocs
+ new WebdavPath.toAbsolutePath() implementation

For toAbsolutePath(), for paths that are absolute paths, it simply return itself.
this conforms with java.nio.file.Path specs.

However, for relative paths, the paths are altered. In 'file://' (unix) implementation,
Path.toAbsolutePath() apparently resolves relative path against the current working directory.

However, with web and WebDAV url it is difficult to define 'current working directory'.
To make this implementation more usable, WebdavPathtoAbsolutePath() use a
'current working path' instance variable to resolve relative paths into absolute paths.
If 'current working path' is not found (null), it is resolved against the default root path "/".

The prefix absolute path are preserved when relative paths are created with several methods.

 - WebdavPath.relativize(Path path)
 - WebdavPath.getFileName()
 - WebdavPath.getName(int index)

To get predictable behavior e.g. that it would be resolved against "current work path"
it is necessary that the relative path is derived by calling the above methods against
base objects that are absolute paths. e.g.

'/a/b/c'.relativize('/a/b') = 'c', '/a/b' preserved as "current work path",
'c'.toAbsolutePath() returns '/a/b/c'

'/a/b/c'.getFileName() = 'c', '/a/b' preserved as "current work path",
'c'.toAbsolutePath() returns '/a/b/c'

'/a/b/c'.getName(1) = 'b', '/a' preserved as "current work path",
'b'.toAbsolutePath() returns '/a/b'

However, it is not saved in the cases as relative paths are used

'a/b/c'.relativize('a/b') = 'c', current work path is not saved as
'a/b/c' is a relative path

hence, 'c'.toAbsolutePath() returns '/c'

the current work path can be retrieved and set using the methods
WebdavPath.getCurrentWorkPath()
WebdavPath.setCurrentWorkPath(WebdavPath path)

However, the currentworkpath needs to be set before calling toAbsolutePath()
and the currentworkpath must not be null to have it resolved that way deterministically

+ updated javadocs for WebdavPath

javadocs comments has been updated for WebdavPath which explained
its usage in detail.